python-api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
Pillow>=10.1.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
